refactor(community): extract useCommunityData + useCommunityActions hooks (751→324 LOC)
Move 14 React Query hooks, 4 side-effects, and 12 handler functions
into dedicated hooks. page.tsx retains only reducer, UI state, and render.

- useCommunityData (140 LOC): queries + derived data + vote/subscription side-effects
- useCommunityActions (331 LOC): all 12 handlers (posts, research, bounties, votes, polls)
- Shared types extracted to hooks/types.ts
- All 55 existing tests pass, 0 tsc errors
